requirements.txt
asgiref==3.3.1
crispy-tailwind==0.4.0
Django==3.1.7
django-crispy-forms==1.11.2
django-phonenumber-field==5.0.0
phonenumbers==8.12.21
psycopg2==2.7.7
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
